move EmailNotification to its own app
